+++ changelog.md
@@ -292,6 +292,8 @@
     //    watcha: fixed large-div scraping
     //    wp: improved some logic (in vain) for non-english titles searching
     //    TODO: make sure edit() work!!!
-    // ver 0.4.87 @ 2023-6-4
-    //    disney+: improved title search...
+    // ver 0.4.89 @ 2023-6-4
+    //    disney+: improved title searching
     //    settings: now numberToBaseEl is generated automatically if omitted.
+    //    watcha: fixed a bug of id searching of v0.4.85
+    //    disney+: fixed edit() behaviors (along with settings)